fix:img size and merge all AcitiyApi to activityApi
